fixed builds and moved another file
